+++ R/bold_seq.R
@@ -3,25 +3,19 @@
 #' Get sequences for a taxonomic name, id, bin, container, institution, 
 #' researcher, geographic, place, or gene.
 #'
-#' @importFrom stringr str_replace_all str_replace str_split
 #' @export
 #' @template args
 #' @template otherargs
+#' @template large-requests
+#' @template marker
 #' @references 
-#' \url{http://v4.boldsystems.org/index.php/resources/api?type=webservices}
+#' http://v4.boldsystems.org/index.php/resources/api?type=webservices
 #'
 #' @param marker (character) Returns all records containing matching 
 #' marker codes.
 #'
-#' @return A list with each element of length 4 with slots for id, name, 
-#' gene, and sequence.
-#' 
-#' @section If a request times out:
-#' This is likely because you're request was for a large number of 
-#' sequences and the BOLD service timed out. You still should get 
-#' some output, those sequences that were retrieved before the time 
-#' out happened. See the README (https://github.com/ropensci/bold/#bold)
-#' for an example of dealing with large data problems with this function.
+#' @return A data frame with each element as row and 5 columns for processid, identification, 
+#' marker, accession, and sequence.
 #' 
 #' @examples \dontrun{
 #' res <- bold_seq(taxon='Coelioxys')
@@ -51,22 +45,21 @@
 bold_seq <- function(taxon = NULL, ids = NULL, bin = NULL, container = NULL, 
   institutions = NULL, researchers = NULL, geo = NULL, marker = NULL, 
   response=FALSE, ...) {
-  
-  args <- bc(
-    list(
-      taxon = pipeornull(taxon), geo = pipeornull(geo), 
-      ids = pipeornull(ids), bin = pipeornull(bin), 
-      container = pipeornull(container), 
-      institutions = pipeornull(institutions),
-      researchers = pipeornull(researchers), marker = pipeornull(marker)
-    )
-  )
-  check_args_given_nonempty(
-    args, 
-    c('taxon','ids','bin','container','institutions','researchers',
-      'geo','marker')
-  )
-  out <- b_GET(paste0(bbase(), 'API_Public/sequence'), args, ...)
+  params <- list(taxon = taxon,
+                  ids = ids,
+                  bin = bin,
+                  container = container,
+                  institutions = institutions,
+                  researchers = researchers,
+                  geo = geo,
+                  marker = marker)
+  paramsNames = names(params)
+  params = params[vapply(params, function(x){is.character(x)&&x!=""}, F)]
+  if(length(params)==0){
+    stop(paste0("You must provide a non-empty value to at least one of\n  ", paste(paramsNames, collapse = "\n  ")))
+  }
+  params <- vapply(params, paste, collapse = "|", "")
+  out <- b_GET(paste0(bbase(), 'API_Public/sequence'), params, ...)
   if (response) { 
     out 
   } else {
@@ -77,6 +70,6 @@ bold_seq <- function(taxon = NULL, ids = NULL, bin = NULL, container = NULL,
       tt <- strdrop(str = tt, pattern = "Fatal+")[[1]]
     }
     res <- strsplit(tt, ">")[[1]][-1]
-    lapply(res, split_fasta)
+    split_fasta(res)
   }
 }