+++ man-roxygen/otherargs.R
@@ -1,5 +1,6 @@
-#' @param response (logical) Note that response is the object that returns 
-#' from the Curl call, useful for debugging, and getting detailed info on 
+#' @param response (logical) Note that response is the object that returns
+#' from the Curl call, useful for debugging, and getting detailed info on
 #' the API call.
-#' @param ... Further args passed on to \code{\link[crul]{HttpClient}}, main 
-#' purpose being curl debugging
+#' @param ... Further args passed on to
+#' \code{\link[crul:verb-GET]{crul::verb-GET}}, main purpose being curl
+#' debugging